Backend/requirements.txt
# Add project dependencies here
pytest
fastapi
uvicorn
sentence-transformers
pypdf2
requests
torch
aiohttp
pdfplumber
nltk 
scikit-learn 
pandas 
numpy
# Dependencies for authentication
motor
pyjwt
passlib[bcrypt]
python-multipart
email-validator
pydantic-settings
google-generativeai
